LAR53-Migrations: last improvements before Merge

+++ modules/BillingBase/Database/Migrations/2016_02_24_100223_create_Product_table.php
@@ -21,8 +21,8 @@ class CreateProductTable extends BaseMigration {
 
 			$table->string('name');
 			$table->enum('type', ['Internet', 'TV', 'Voip', 'Device', 'Credit', 'Other']);
-			$table->integer('qos_id')->unsigned()->nullable();
-			$table->integer('voip_id')->unsigned()->nullable();
+			$table->tinyInteger('qos_id')->unsigned()->nullable();
+			$table->tinyInteger('voip_id')->unsigned()->nullable();
 			$table->enum('billing_cycle', ['Once', 'Monthly', 'Quarterly', 'Yearly']);
 			$table->integer('costcenter_id')->unsigned();
 			$table->float('price');
